% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installers.R
\name{installers}
\alias{installers}
\alias{install.state}
\alias{install.county}
\alias{install.tract}
\alias{install.cdp}
\alias{install.blkgrp}
\alias{install.block}
\title{Data package installers}
\usage{
install.state()

install.county()

install.tract()

install.cdp()

install.blkgrp()

install.block()
}
\value{
nothing, called to install a data package
}
\description{
Data package installers
}
